+++ packages/docs/src/stories/TextArea.stories.tsx
@@ -1,6 +1,6 @@
 import type { StoryObj, Meta } from '@storybook/react'
 
-import { Box, Text, TextArea, TextAreaProps } from '@alosix-ui/react'
+import { Box, Text, TextArea, TextAreaProps } from '@alosix-hub-ui/react'
 
 export default {
   title: 'React/Form/Text Area',