+++ packages/tokens/src/colors.ts
@@ -16,5 +16,5 @@ export const colors = {
   alosix700: '#015F43',
   alosix900: '#00291D',
 
-  teste: '#f00'
+  teste: '#f02'
 }