+++ packages/eslint-config/CHANGELOG.md
@@ -1,4 +1,10 @@
-# @alosix-ui/eslint-config
+# @alosix-hub-ui/eslint-config
+
+## 3.0.0
+
+### Major Changes
+
+- Core
 
 ## 2.0.0